+++ components/RecurringTasks/streaks/patterns/daily/types.ts
@@ -1,0 +1,6 @@
+import { DateRange } from '../../types';
+
+export interface DailyTarget {
+  date: Date;
+  allowedRange: DateRange;
+}
